Milestone 2
The calculate method no longer just returns 10.0 to meet the basic test
function. Instead it has a more comprehensive calculate function to
determine the percentage discount on a given total.

The test includes fifteen_percent_discount. This sserts that the method
in discount_calculator returns the float of 15.0 when given the total of
100 and with the percent of 15.