Milestone 3
Adds the five_dollar_discount test which asserts that the result is $5.
This is clearly a fail as the input is $250 and the 5% discount for that
is $12.50. This doesn't work.

We reword the calculate() function with an if/else statement. If the
discount_type is anything other than a 'percent', we treat it as an
absolute and the discount is equivalent to the inputted discount_amount.