Milestone 5
We've added two new tests. The first one asserts that the result of the
float of the input 100.0 and percentage of 10.0 will be 10.0.

The second asserts that the absolute value of $250 with $5.00 as the
discount value will be $5.

Both tests pass on first try so no refactoring needs to be done. Still,
it's always good to reinforce your code to ensure no errors.